Create card-status-dropdown and move the messy thing into there
